zheqi - move functions in utils/fragment_utility.py to modules/fragment/reduce_hydrogen.py and delete utils/fragment_utility.py
